# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ArcLengthSample": {
        "properties": {
          "distance": {
            "title": "Distance",
            "type": "number"
          },
          "segment_index": {
            "title": "Segment Index",
            "type": "integer"
          },
          "t": {
            "title": "T",
            "type": "number"
          }
        },
        "required": [
          "segment_index",
          "t",
          "distance"
        ],
        "title": "ArcLengthSample",
        "type": "object"
      },
...
      "CameraOrientation": {
        "properties": {
          "pitch_deg": {
            "default": 0.0,
            "title": "Pitch Deg",
            "type": "number"
          },
          "roll_deg": {
            "default": 0.0,
            "title": "Roll Deg",
            "type": "number"
          },
          "yaw_deg": {
            "default": 0.0,
            "title": "Yaw Deg",
            "type": "number"
          }
        },
        "title": "CameraOrientation",
        "type": "object"
      },
      "CameraOrientationKeyframe": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interpolation_to_next": {
            "default": "smoothstep",
            "enum": [
              "smoothstep",
              "linear",
              "hold"
            ],
            "title": "Interpolation To Next",
            "type": "string"
          },
          "orientation": {
            "$ref": "#/components/schemas/CameraOrientation"
          },
          "path_position": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Path Position",
            "type": "number"
          }
        },
        "required": [
          "path_position",
          "orientation"
        ],
        "title": "CameraOrientationKeyframe",
        "type": "object"
      },
...
      "CenterWeightedDepthOfFieldFocus": {
        "properties": {
          "kind": {
            "const": "center_weighted_9",
            "default": "center_weighted_9",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "CenterWeightedDepthOfFieldFocus",
        "type": "object"
      },
      "ChatHistoryMessage": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatHistoryMessage",
        "type": "object"
      },
      "ChatMessage": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "compiled": {
            "$ref": "#/components/schemas/CompiledTrajectory"
          }
        },
        "required": [
          "answer",
          "compiled"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "CompiledCameraKeyframe": {
        "properties": {
          "aim": {
            "discriminator": {
              "mapping": {
                "follow_path": "#/components/schemas/FollowPathAim",
                "look_at_point": "#/components/schemas/ResolvedLookAtPointAim"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FollowPathAim"
              },
              {
                "$ref": "#/components/schemas/ResolvedLookAtPointAim"
              }
            ],
            "title": "Aim"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interpolation_to_next": {
            "enum": [
              "smoothstep",
              "linear",
              "hold"
            ],
            "title": "Interpolation To Next",
            "type": "string"
          },
          "path_position": {
            "title": "Path Position",
            "type": "number"
          }
        },
        "required": [
          "id",
          "path_position",
          "aim",
          "interpolation_to_next"
        ],
        "title": "CompiledCameraKeyframe",
        "type": "object"
      },
      "CompiledCameraTrack": {
        "properties": {
          "default_aim": {
            "discriminator": {
              "mapping": {
                "follow_path": "#/components/schemas/FollowPathAim",
                "look_at_point": "#/components/schemas/ResolvedLookAtPointAim"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FollowPathAim"
              },
              {
                "$ref": "#/components/schemas/ResolvedLookAtPointAim"
              }
            ],
            "title": "Default Aim"
          },
          "default_orientation": {
            "$ref": "#/components/schemas/CameraOrientation"
          },
          "depth_of_field_keyframes": {
            "items": {
              "$ref": "#/components/schemas/CompiledDepthOfFieldKeyframe"
            },
            "title": "Depth Of Field Keyframes",
            "type": "array"
          },
          "keyframes": {
            "items": {
              "$ref": "#/components/schemas/CompiledCameraKeyframe"
            },
            "title": "Keyframes",
            "type": "array"
          },
          "orientation_keyframes": {
            "items": {
              "$ref": "#/components/schemas/CameraOrientationKeyframe"
            },
            "title": "Orientation Keyframes",
            "type": "array"
          },
          "world_up": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "World Up",
            "type": "array"
          }
        },
        "required": [
          "default_aim",
          "keyframes",
          "default_orientation",
          "orientation_keyframes",
          "depth_of_field_keyframes",
          "world_up"
        ],
        "title": "CompiledCameraTrack",
        "type": "object"
      },
      "CompiledDepthOfFieldKeyframe": {
        "properties": {
          "bokeh_scale": {
            "title": "Bokeh Scale",
            "type": "number"
          },
          "focus": {
            "discriminator": {
              "mapping": {
                "center_weighted_9": "#/components/schemas/CenterWeightedDepthOfFieldFocus",
                "scene_point": "#/components/schemas/ResolvedScenePointDepthOfFieldFocus"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/CenterWeightedDepthOfFieldFocus"
              },
              {
                "$ref": "#/components/schemas/ResolvedScenePointDepthOfFieldFocus"
              }
            ],
            "title": "Focus"
          },
          "focus_range_scale": {
            "title": "Focus Range Scale",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "path_position": {
            "title": "Path Position",
            "type": "number"
          }
        },
        "required": [
          "id",
          "path_position",
          "focus",
          "focus_range_scale",
          "bokeh_scale"
        ],
        "title": "CompiledDepthOfFieldKeyframe",
        "type": "object"
      },
      "CompiledMotionProfile": {
        "properties": {
          "default_speed": {
            "title": "Default Speed",
            "type": "number"
          },
          "keyframes": {
            "items": {
              "$ref": "#/components/schemas/SpeedKeyframe"
            },
            "title": "Keyframes",
            "type": "array"
          }
        },
        "required": [
          "default_speed",
          "keyframes"
        ],
        "title": "CompiledMotionProfile",
        "type": "object"
      },
      "CompiledTrajectory": {
        "properties": {
          "arc_length_table": {
            "items": {
              "$ref": "#/components/schemas/ArcLengthSample"
            },
            "title": "Arc Length Table",
            "type": "array"
          },
          "camera_track": {
            "$ref": "#/components/schemas/CompiledCameraTrack"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "number"
          },
          "motion_profile": {
            "$ref": "#/components/schemas/CompiledMotionProfile"
          },
          "position_segments": {
            "items": {
              "$ref": "#/components/schemas/CubicBezier3D"
            },
            "title": "Position Segments",
            "type": "array"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "total_length": {
            "title": "Total Length",
            "type": "number"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "project_id",
          "revision",
          "position_segments",
          "arc_length_table",
          "total_length",
          "duration_seconds",
          "motion_profile",
          "camera_track"
        ],
        "title": "CompiledTrajectory",
        "type": "object"
      },
      "CubicBezier3D": {
        "properties": {
          "length": {
            "title": "Length",
            "type": "number"
          },
          "p0": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "P0",
            "type": "array"
          },
          "p1": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "P1",
            "type": "array"
          },
          "p2": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "P2",
            "type": "array"
          },
          "p3": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "P3",
            "type": "array"
          },
          "source_segment_id": {
            "title": "Source Segment Id",
            "type": "string"
          }
        },
        "required": [
          "source_segment_id",
          "p0",
          "p1",
          "p2",
          "p3",
          "length"
        ],
        "title": "CubicBezier3D",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "FollowPathAim": {
        "properties": {
          "direction": {
            "default": "forward",
            "enum": [
              "forward",
              "backward"
            ],
            "title": "Direction",
            "type": "string"
          },
          "kind": {
            "const": "follow_path",
            "default": "follow_path",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "FollowPathAim",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolvedLookAtPointAim": {
        "properties": {
          "kind": {
            "const": "look_at_point",
            "default": "look_at_point",
            "title": "Kind",
            "type": "string"
          },
          "position": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Position",
            "type": "array"
          },
          "scene_point_id": {
            "title": "Scene Point Id",
            "type": "string"
          }
        },
        "required": [
          "scene_point_id",
          "position"
        ],
        "title": "ResolvedLookAtPointAim",
        "type": "object"
      },
      "ResolvedScenePointDepthOfFieldFocus": {
        "properties": {
          "kind": {
            "const": "scene_point",
            "default": "scene_point",
            "title": "Kind",
            "type": "string"
          },
          "position": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Position",
            "type": "array"
          },
          "scene_point_id": {
            "title": "Scene Point Id",
            "type": "string"
          }
        },
        "required": [
          "scene_point_id",
          "position"
        ],
        "title": "ResolvedScenePointDepthOfFieldFocus",
        "type": "object"
      },
...
      "SpeedKeyframe": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interpolation_to_next": {
            "default": "smoothstep",
            "enum": [
              "smoothstep",
              "linear",
              "hold"
            ],
            "title": "Interpolation To Next",
            "type": "string"
          },
          "path_position": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Path Position",
            "type": "number"
          },
          "speed": {
            "exclusiveMinimum": 0.0,
            "title": "Speed",
            "type": "number"
          }
        },
        "required": [
          "path_position",
          "speed"
        ],
        "title": "SpeedKeyframe",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Create camera-path projects, author scene and trajectory data, edit playback timelines, and run the trajectory agent. Canonical operations use project revision ETags for optimistic concurrency.",
    "summary": "Semantic 3D camera trajectory authoring API",
    "title": "Camera Path API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/chat": {
      "delete": {
        "description": "Delete all chat messages for the project.",
        "operationId": "clearProjectChat",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found.",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear chat history",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/projects/{project_id}/chat/messages": {
      "get": {
        "description": "Return the complete chat history for the first API version.",
        "operationId": "listChatMessages",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatHistoryMessage"
                  },
                  "title": "Response Listchatmessages",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List chat messages",
        "tags": [
          "Chat"
        ]
      },
      "post": {
        "description": "Run the agent and return its answer plus derived compiled trajectory.",
        "operationId": "createChatMessage",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send a chat message",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/projects/{project_id}/chat/messages/stream": {
      "post": {
        "description": "Run the agent as SSE. `delta` events contain text, `result` contains the final answer and compiled trajectory without a full Project snapshot, and `error` reports failure.",
        "operationId": "streamChatMessage",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent events named delta, result, or error."
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream a chat response",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/projects/{project_id}/chat/user-messages": {
      "post": {
        "description": "Persist and return one idempotent user message.",
        "operationId": "saveUserChatMessage",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryMessage"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save a user chat message",
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
...
}